fix: match espShow's pin parameter type across translation units
Adafruit_NeoPixel.cpp declares (and calls) the ESP8266/ESP32 espShow()
helper with a uint16_t pin parameter, matching the class's own
`int16_t pin` member:

  extern "C" void espShow(uint16_t pin, uint8_t *pixels, uint32_t numBytes, uint8_t type);

but esp.c and esp8266.c both define it with uint8_t pin instead - a
real cross-TU type mismatch (a One Definition Rule violation for an
extern "C" function). Building with -flto correctly flags this:

  warning: type of 'espShow' does not match original declaration [-Wlto-type-mismatch]
  note: type 'uint8_t' should match type 'uint16_t'

Most builds don't use LTO so this goes unnoticed in practice, but it's
still undefined behavior per the standard, and nothing guarantees the
mismatch stays harmless across compilers/optimization levels.

Fixed esp.c's two espShow() definitions and esp8266.c's two
(ESP8266/non-ESP8266) definitions to use uint16_t pin, matching the
declaration.

Fixes #400

+++ esp.c
@@ -35,7 +35,7 @@
 
 static SemaphoreHandle_t show_mutex = NULL;
 
-void espShow(uint8_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
+void espShow(uint16_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
   // Note: Because rmtPin is shared between all instances, we will
   //  end up releasing/initializing the RMT channels each time we
   //  invoke on different pins. This is probably ok, just not
@@ -185,7 +185,7 @@ static void IRAM_ATTR ws2812_rmt_adapter(const void *src, rmt_item32_t *dest, si
     *item_num = num;
 }
 
-void espShow(uint8_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
+void espShow(uint16_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
     // Reserve channel
     rmt_channel_t channel = ADAFRUIT_RMT_CHANNEL_MAX;
     for (size_t i = 0; i < ADAFRUIT_RMT_CHANNEL_MAX; i++) {
